Initial PE component implementation
