+++ Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 89343fd38ebd2004697150d1be60b070
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
