+++ Scripts/Editor/BUDDYWORKS.PosesExtension.Component.Banner.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 01833208bc0486741ac585c33c8e9f50
+timeCreated: 1740014242
