+++ Scripts/Editor/BUDDYWORKS.PosesExtension.Component.Editor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: cade6287cbb377c49ad38c31c779984f
+timeCreated: 1740014242
