+++ Scripts/Editor/BUDDYWORKS.PosesExtension.Component.Strings.cs
@@ -1,0 +1,15 @@
+﻿// Collection of reused strings across scripts.
+
+namespace BUDDYWORKS.PosesExtension
+{
+    public class BWStrings
+    {
+        public static string logHeader = "<color=grey>[</color><color=#FDDA0D>BUDDYWORKS</color><color=grey>] </color>";
+        public static string logAbort = logHeader + "<color=red>ERROR:</color> ";
+        public static string logInfo = logHeader + "<color=white>Log:</color> ";
+        public static string logSuccess = logHeader + "<color=green>OK:</color> ";
+        
+        public static string bwAOC = "<color=grey>[</color><color=white>AO</color><color=grey>]</color>";
+        public static string bwUB = "<color=grey>[</color><color=white>UB</color><color=grey>]</color>";
+    }
+}
