+++ Scripts/Editor/BUDDYWORKS.PosesExtension.Component.Strings.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 423fe5160f3b2c54eb504480c8fc39f1
+timeCreated: 1740014242
