+++ Scripts/Runtime/BWPosesExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df58f6a6e612d234396cce2b08210871
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 25cfaf88af7b1fe45b026df2ac43a052, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
