+++ Scripts/wtf.buddyworks.posesextension.asmdef
@@ -2,9 +2,7 @@
     "name": "wtf.buddyworks.posesextension",
     "rootNamespace": "",
     "references": [],
-    "includePlatforms": [
-        "Editor"
-    ],
+    "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
     "overrideReferences": false,
